+++ sdk/core/core-https/src/policies/bearerTokenAuthenticationPolicy.ts
@@ -29,9 +29,39 @@ export interface BearerTokenAuthenticationPolicyOptions {
    */
   scopes: string | string[];
   /**
-   * Allows the dynamic discovery of authentication properties.
+   * Allows for the processing of [Continuous Access Evaluation](https://docs.microsoft.com/azure/active-directory/conditional-access/concept-continuous-access-evaluation) challenges.
+   * If provided, it must contain at least the `processChallenge` method.
+   * If provided, after a request is sent, if it has a challenge, it can be processed to re-send the original request with the relevant challenge information.
    */
-  authenticationContext?: AuthenticationContext;
+  challenge?: {
+    /**
+     * Allows for the customization of the next request before its sent.
+     * By default we won't be doing any changes to the initial challenge request.
+     */
+    prepareRequest?(request: PipelineRequest): Promise<void>;
+    /**
+     * Defines how to get the challenge from the PipelineResponse.
+     * By default we will retrieve the challenge only if the response status code was 401,
+     * and if the response contained the header "WWW-Authenticate" with a non-empty value.
+     */
+    getChallenge?(response: PipelineResponse): string | undefined;
+    /**
+     * Updates  the authentication context based on the challenge.
+     */
+    processChallenge(challenge: string): Promise<AuthenticationContext | undefined>;
+  };
+}
+
+/**
+ * By default we will retrieve the challenge only if the response status code was 401,
+ * and if the response contained the header "WWW-Authenticate" with a non-empty value.
+ */
+function defaultGetChallenge(response: PipelineResponse): string | undefined {
+  const challenges = response.headers.get("WWW-Authenticate");
+  if (response.status === 401 && challenges) {
+    return challenges;
+  }
+  return;
 }
 
 /**
@@ -43,32 +73,73 @@ export function bearerTokenAuthenticationPolicy(
 ): PipelinePolicy {
   const { credential, scopes } = options;
   const tokenCache: AccessTokenCache = new ExpiringAccessTokenCache();
-  const authenticationContext: AuthenticationContext = options.authenticationContext || {};
+  const { prepareRequest, getChallenge = defaultGetChallenge, processChallenge } =
+    options.challenge || {};
 
-  async function getToken(tokenOptions: GetTokenOptions): Promise<string | undefined> {
+  /**
+   * getToken will call to the underlying credential's getToken request with properties coming from the request,
+   * as well as properties coming from parsing the CAE challenge, if any.
+   */
+  async function getToken(
+    request: PipelineRequest,
+    context?: AuthenticationContext
+  ): Promise<string | undefined> {
     let accessToken = tokenCache.getCachedToken();
+    const getTokenOptions: GetTokenOptions = {
+      claims: context?.challengeClaims,
+      abortSignal: request.abortSignal,
+      tracingOptions: {
+        spanOptions: request.spanOptions
+      }
+    };
     if (accessToken === undefined) {
       accessToken =
-        (await credential.getToken(authenticationContext.scopes || scopes, tokenOptions)) ||
-        undefined;
+        (await credential.getToken(context?.scopes || scopes, getTokenOptions)) || undefined;
       tokenCache.setCachedToken(accessToken);
     }
-
     return accessToken ? accessToken.token : undefined;
   }
+
+  /**
+   * Populates the token in the request headers.
+   */
+  function assignToken(request: PipelineRequest, token?: string): PipelineRequest {
+    request.headers.set("Authorization", `Bearer ${token}`);
+    return request;
+  }
+
+  /**
+   * Uses the challenge parameters to:
+   * - Prepare the outgoing request (if the `prepareRequest` method has been provided).
+   * - Process a challenge if the response contains it.
+   * - Retrieve a token with the challenge information, then re-send the request.
+   */
+  async function challengePolicy(
+    request: PipelineRequest,
+    next: SendRequest
+  ): Promise<PipelineResponse> {
+    if (prepareRequest) {
+      await prepareRequest(request);
+    }
+    const response = await next(request);
+    const challenge = getChallenge(response);
+    if (challenge && processChallenge) {
+      const context = await processChallenge(challenge);
+      const token = await getToken(request, context);
+      return next(assignToken(request, token));
+    }
+    return response;
+  }
+
   return {
     name: bearerTokenAuthenticationPolicyName,
     async sendRequest(request: PipelineRequest, next: SendRequest): Promise<PipelineResponse> {
-      const token = await getToken({
-        abortSignal: request.abortSignal,
-        tracingOptions: {
-          spanOptions: request.spanOptions
-        },
-        claims: authenticationContext.claims
-      });
-
-      request.headers.set("Authorization", `Bearer ${token}`);
-      return next(request);
+      const token = await getToken(request);
+      if (token) {
+        return next(assignToken(request, token));
+      } else {
+        return await challengePolicy(request, next);
+      }
     }
   };
 }